+++ test/browser/helpers/ssr-components.js
@@ -35,5 +35,35 @@ export class SsrRoot extends Symbiote {
 SsrRoot.template = html`
   <h1 ${{textContent: 'rootTitle'}}></h1>
   <ssr-middle></ssr-middle>
+  <ssr-inner-html></ssr-inner-html>
+  <ssr-list></ssr-list>
 `;
 SsrRoot.reg('ssr-root');
+
+export class SsrInnerHtml extends Symbiote {
+  ssrMode = true;
+  init$ = {
+    richContent: '<em>server rendered</em>',
+  };
+}
+SsrInnerHtml.template = html`<div ${{innerHTML: 'richContent'}}></div>`;
+SsrInnerHtml.reg('ssr-inner-html');
+
+export class SsrList extends Symbiote {
+  ssrMode = true;
+  init$ = {
+    items: [
+      { label: 'Item A' },
+      { label: 'Item B' },
+      { label: 'Item C' },
+    ],
+  };
+}
+SsrList.template = html`
+  <ul itemize="items">
+    <template>
+      <li ${{textContent: 'label'}}></li>
+    </template>
+  </ul>
+`;
+SsrList.reg('ssr-list');